Fix migrations Fix id's Rename DiscordServer to Server

+++ Eventee/Entities/EventMember.cs
@@ -3,7 +3,7 @@ namespace Eventee.Entities;
 public class EventMember
 {
     public ulong MemberId { get; set; }
-    public ulong EventId { get; set; }
+    public int EventId { get; set; }
     
     public Member Member { get; set; }
     public Event Event { get; set; }
